# backend/openapi/unital-v1.yaml
openapi: 3.1.0
info:
  title: Unital API
  version: 1.0.0
  description: |
    Building & complex management platform. Consolidated gateway API for the
    Unital Go microservices (identity, property, billing, facilities,
    operations, notifications/communications).

    **Auth**: `POST /auth/login` returns an access token (15 min) + refresh
    token. Send `Authorization: Bearer <access>` on every other call.

    **Errors**: all failures use RFC 9457 problem details with a
    machine-readable `code` and `trace_id`.
  contact: { name: Unital, url: https://unital.app }
servers:
  - url: http://localhost:8080/api/v1
    description: local gateway
  - url: http://localhost:18080/api/v1
    description: local gateway (dev port)
...
security:
  - bearerAuth: []
paths:
...
  /facilities:
    post:
      tags: [Facilities]
      summary: Define a facility (manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FacilityInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Facility" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /buildings/{buildingID}/facilities:
    parameters:
      - $ref: "#/components/parameters/BuildingID"
    get:
      tags: [Facilities]
      summary: List the building's facilities (members)
      responses:
        "200":
          description: Facilities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Facility" }
  /facilities/{facilityID}:
    parameters:
      - name: facilityID
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Facilities]
      summary: Facility detail
      responses:
        "200":
          description: Facility
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Facility" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [Facilities]
      summary: Update a facility (manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FacilityInput" }
      responses:
        "200": { description: Updated facility }
        "403": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [Facilities]
      summary: Delete a facility (manager)
      responses:
        "200": { description: Deleted }
        "403": { $ref: "#/components/responses/Problem" }
  /facilities/{facilityID}/availability:
    parameters:
      - name: facilityID
        in: path
        required: true
        schema: { type: string }
      - name: date
        in: query
        schema: { type: string, format: date }
    get:
      tags: [Facilities]
      summary: Free hours for a date (defaults to today)
      responses:
        "200":
          description: Free hours of day
          content:
            application/json:
              schema:
                type: object
                properties:
                  date: { type: string, format: date }
                  free_hours:
                    type: array
                    items: { type: integer }
  /facilities/{facilityID}/bookings:
    post:
      tags: [Facilities]
      summary: Book a facility (building members; slot rules enforced)
      parameters:
        - name: facilityID
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingInput" }
      responses:
        "201":
          description: Booking created (pending approval)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /bookings:
    get:
      tags: [Facilities]
      summary: My bookings
      parameters:
        - name: mine
          in: query
          required: true
          schema: { type: string, enum: ["true"] }
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Booking" }
  /bookings/{bookingID}:
    get:
      tags: [Facilities]
      summary: Booking detail
      parameters:
        - name: bookingID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "404": { $ref: "#/components/responses/Problem" }
  /bookings/{bookingID}/approve:
    post:
      tags: [Facilities]
      summary: Approve a pending booking (manager; conflicts rejected)
      parameters:
        - name: bookingID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Confirmed booking }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /bookings/{bookingID}/reject:
    post:
      tags: [Facilities]
      summary: Reject a pending booking (manager)
      parameters:
        - name: bookingID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Rejected booking }
        "403": { $ref: "#/components/responses/Problem" }
  /bookings/{bookingID}/cancel:
    post:
      tags: [Facilities]
      summary: Cancel a booking (owner or manager)
      parameters:
        - name: bookingID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Cancelled booking }
        "403": { $ref: "#/components/responses/Problem" }
  /facilities/{facilityID}/maintenance-windows:
    post:
      tags: [Facilities]
      summary: Schedule maintenance (manager; cancels overlapping bookings)
      parameters:
        - name: facilityID
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, scheduled_start, scheduled_end]
              properties:
                title: { type: string }
                type: { type: string, enum: [routine, repair, cleaning, inspection, upgrade] }
                priority: { type: string, enum: [low, medium, high, urgent] }
                scheduled_start: { type: string, format: date-time }
                scheduled_end: { type: string, format: date-time }
                affect_bookings: { type: boolean, default: true }
      responses:
        "201": { description: Maintenance window created }
        "403": { $ref: "#/components/responses/Problem" }
  /maintenance-windows/{maintenanceID}/start:
    post:
      tags: [Facilities]
      summary: Start scheduled maintenance (manager)
      parameters:
        - name: maintenanceID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Maintenance in progress }
        "422": { $ref: "#/components/responses/Problem" }
  /maintenance-windows/{maintenanceID}/complete:
    post:
      tags: [Facilities]
      summary: Complete maintenance (manager)
      parameters:
        - name: maintenanceID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Maintenance done }
        "422": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BuildingID:
      name: buildingID
      in: path
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details with machine-readable code
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string, example: UNIT_NOT_IN_BUILDING }
        detail: { type: string }
        trace_id: { type: string }
...
    FacilityInput:
      type: object
      required: [building_id, name, capacity]
      properties:
        building_id: { type: string }
        name: { type: string }
        type: { type: string, enum: [pool, gym, roof_garden, meeting_room, party_hall, guest_parking, playground, sports_court, library, business_center, other] }
        description: { type: string }
        capacity: { type: integer, minimum: 1 }
        opening_time: { type: string, example: "08:00" }
        closing_time: { type: string, example: "22:00" }
        hourly_rate: { type: integer, description: Rials per hour (0 = free) }
        min_advance_hours: { type: integer, default: 1 }
        max_advance_hours: { type: integer, default: 168 }
        rules: { type: array, items: { type: string } }
        images: { type: array, items: { type: string } }
    Facility:
      allOf:
        - $ref: "#/components/schemas/FacilityInput"
        - type: object
          properties:
            id: { type: string }
            is_active: { type: boolean }
            created_at: { type: string, format: date-time }
    BookingInput:
      type: object
      required: [start, end]
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        purpose: { type: string }
        participants: { type: integer, minimum: 1 }
    Booking:
      type: object
      properties:
        id: { type: string }
        facility_id: { type: string }
        building_id: { type: string }
        user_id: { type: string }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        purpose: { type: string }
        participants: { type: integer }
        status: { type: string, enum: [pending, confirmed, completed, cancelled, rejected] }
        total_cost: { type: integer }
        approved_by: { type: string }
        created_at: { type: string, format: date-time }
...